+++ package.json
@@ -37,6 +37,7 @@
   "devDependencies": {
     "@types/node": "^22.15.21",
     "esbuild": "^0.25.4",
+    "execa": "^9.6.0",
     "vitest": "^3.1.4"
   }
 }
