+++ src/utils.ts
@@ -1,0 +1,8 @@
+import fs from 'fs';
+
+export const getPkgManager = () => {
+    if (fs.existsSync('pnpm-lock.yaml')) return 'pnpm';
+    if (fs.existsSync('yarn.lock')) return 'yarn';
+
+    return 'npm';
+};
